Fix sporadic failures due to taken port
