+++ src/grpcservermodule.hpp
@@ -51,8 +51,5 @@ public:
 
     const GetModelMetadataImpl& getTFSModelMetadataImpl() const;
     KFSInferenceServiceImpl& getKFSGrpcImpl() const;
-
-private:
-    bool isPortAvailable(uint64_t port);
 };
 }  // namespace ovms
